requirements.txt
Pillow
pyserial
PIL
